requirements.txt
Flask
opencv-python
numpy
Pillow
timm
torch
torchvision
werkzeug
fastapi
uvicorn
pydantic
python-multipart
